feat: Properly show broken bonds on GPU
- Add Inactive bonds counter.
- Distinguish between inactive and broken bonds on GPU.
- Remove liquid bonds counter.
- Update simulation statistics even when no PP and PW models selected.
- closes #31

+++ Models/SolidBonds/BondModelElasticPerfectlyPlastic/ModelSBElasticPerfectlyPlastic.cu
@@ -78,7 +78,7 @@ __global__ void CUDA_CalcSBForce_EPP_kernel(
 	const double	_bondTangentialStrengths[],
 	const double	_bondYieldStrengths[],
 
-	uint8_t		_bondActivities[],
+	unsigned	_bondActivities[],
 	double		_bondEndActivities[],
 	CVector3	_bondNormalMoments[],
 	double		_bondNormalPlasticStrains[],
